refactor: Rule 디렉터리 분리
