+++ src/main/java/racingcar/domain/rule/MoveRule.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.rule;
 
 public interface MoveRule {
 	boolean isMovable(int number);
